Bot for harp song Geothermy .1(read desc)
Bot for song Geothermy. THIS IS NOT THE FIRST SONG IN GAME BUT I INDEX IT AS 1. ALL THE OTHER ONES WILL BE INDEXED AFTER THIS.
Use minecraft on smallscreen with it being on the middle of the monitor(default) at 1080px res. I made it dependant on timing so you have to trial and error that.
Dependecy: pyautogui